voltasnaroda funcional

+++ API_PostgreSql/Domain/DTOs/CageDTO.cs
@@ -6,7 +6,7 @@ namespace API_PostgreSql.Domain.DTOs
     public class CageDTO
     {
         public int Id { get; private set; }
-        public int Diametro { get; set; }
+        public float Diametro { get; set; }
         public string Descricao { get; set; }
     }
 }
